+++ lld/lib/ReaderWriter/ELF/PPCReference.cpp
@@ -46,7 +46,7 @@ int relocB24PCREL(uint8_t *location, uint64_t P, uint64_t S, uint64_t A) {
   if ((result < 0x1000000) && (result > -0x1000000)) {
     result &= ~-(0x1000000);
     *reinterpret_cast<llvm::support::ubig32_t *>(location) = result |
-                      *reinterpret_cast<llvm::support::ubig32_t *>(location);
+               (uint32_t)*reinterpret_cast<llvm::support::ubig32_t *>(location);
     return PPCKindHandler::NoError;
   }
   return PPCKindHandler::Overflow;
