+++ lld/lib/ReaderWriter/ELF/X86Reference.cpp
@@ -45,14 +45,14 @@ int relocNone(uint8_t *location, uint64_t P, uint64_t S, uint64_t A) {
 int reloc32(uint8_t *location, uint64_t P, uint64_t S, uint64_t A) {
   int32_t result = (uint32_t)(S + A);
   *reinterpret_cast<llvm::support::ulittle32_t *>(location) = result |
-                    *reinterpret_cast<llvm::support::ulittle32_t *>(location);
+            (uint32_t)*reinterpret_cast<llvm::support::ulittle32_t *>(location);
   return X86KindHandler::NoError;
 }
 /// \brief R_386_PC32 - word32: S + A - P
 int relocPC32(uint8_t *location, uint64_t P, uint64_t S, uint64_t A) {
   uint32_t result = (uint32_t)((S + A) - P);
   *reinterpret_cast<llvm::support::ulittle32_t *>(location) = result +
-                    *reinterpret_cast<llvm::support::ulittle32_t *>(location);
+            (uint32_t)*reinterpret_cast<llvm::support::ulittle32_t *>(location);
   return X86KindHandler::NoError;
 }
